Split form action to keep action operation pattern
