+++ src/app/(auth-pages)/sign-in/page.tsx
@@ -1,16 +1,14 @@
-import { revalidateTag } from 'next/cache'
 import { headers } from 'next/headers'
 import Link from 'next/link'
 import { redirect } from 'next/navigation'
 
-import { authActions } from '@/app/_actions/auth-actions'
 import Field from '@/components/form/field'
 import { FormMessage, Message } from '@/components/form/form-message'
 import { SubmitButton } from '@/components/submit-button'
 import { Input } from '@/components/ui/input'
 import { getAuthUserId } from '@/utils/auth'
-import { encodedRedirect } from '@/utils/encoded-redirect'
-import { tryCatch } from '@/utils/try-catch'
+
+import { signInFormAction } from './signin-form-action'
 
 export default async function Login(props: { searchParams: Promise<Message> }) {
   // If user is already logged in, they don't need to be here.
@@ -52,24 +50,4 @@ export default async function Login(props: { searchParams: Promise<Message> }) {
       </form>
     </>
   )
-}
-
-async function signInFormAction(formData: FormData) {
-  'use server'
-  const email = formData.get('email') as string
-  const password = formData.get('password') as string
-  const redirectTo = formData.get('redirectTo')?.toString() || '/feed'
-
-  const { data: authUser, error } = await tryCatch(authActions.signIn({ email, password }))
-
-  if (error) {
-    return encodedRedirect('error', '/sign-in', error.message)
-  }
-
-  // Revalidate the user cache on successful sign in
-  if (authUser) {
-    revalidateTag(`user-${authUser.id}`)
-  }
-
-  return redirect(redirectTo)
 }
